+++ src/auth/mod.rs
@@ -2,5 +2,7 @@
 
 #[cfg(feature = "approle")]
 pub mod approle;
+#[cfg(feature = "kubernetes-auth")]
+pub mod kubernetes;
 #[cfg(feature = "token")]
 pub mod token;